statistics & plots

+++ SOPS/Controllers/ExistingProductController.cs
@@ -8,7 +8,10 @@ using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Description;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.Owin;
 using SOPS.Attributes;
+using SOPS.ModelHelpers;
 using SOPS.Models;
 
 namespace SOPS.Controllers
@@ -17,6 +20,8 @@ namespace SOPS.Controllers
     public class ExistingProductController : ApiController
     {
         private ApplicationDbContext db = new ApplicationDbContext();
+        public ApplicationRoleManager RoleManager => Request.GetOwinContext().GetUserManager<ApplicationRoleManager>();
+        public ApplicationUserManager UserManager => Request.GetOwinContext().GetUserManager<ApplicationUserManager>();
 
         // GET: api/ExistingProducts
         public IQueryable<ExistingProduct> GetExistingProducts()
@@ -38,6 +43,7 @@ namespace SOPS.Controllers
         }
 
         // PUT: api/ExistingProducts/5
+        [Authorize(Roles = "Employee, Administrator")]
         [ResponseType(typeof(void))]
         public IHttpActionResult PutExistingProduct(int id, ExistingProduct existingProduct)
         {
@@ -49,6 +55,11 @@ namespace SOPS.Controllers
             if (id != existingProduct.Id)
             {
                 return BadRequest();
+            }
+
+            if (!db.IsCurrentUserEmployedInCompanyOrAdministrator(existingProduct.Product.CompanyId))
+            {
+                return StatusCode(HttpStatusCode.Forbidden);
             }
 
             db.Entry(existingProduct).State = EntityState.Modified;
@@ -73,12 +84,18 @@ namespace SOPS.Controllers
         }
 
         // POST: api/ExistingProducts
+        [Authorize(Roles = "Employee, Administrator")]
         [ResponseType(typeof(ExistingProduct))]
         public IHttpActionResult PostExistingProduct(ExistingProduct existingProduct)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
+            }
+
+            if (!db.IsCurrentUserEmployedInCompanyOrAdministrator(existingProduct.Product.CompanyId))
+            {
+                return StatusCode(HttpStatusCode.Forbidden);
             }
 
             db.ExistingProducts.Add(existingProduct);
@@ -88,6 +105,7 @@ namespace SOPS.Controllers
         }
 
         // DELETE: api/ExistingProducts/5
+        [Authorize(Roles = "Employee, Administrator")]
         [ResponseType(typeof(ExistingProduct))]
         public IHttpActionResult DeleteExistingProduct(int id)
         {
@@ -95,6 +113,11 @@ namespace SOPS.Controllers
             if (existingProduct == null)
             {
                 return NotFound();
+            }
+
+            if (!db.IsCurrentUserEmployedInCompanyOrAdministrator(existingProduct.Product.CompanyId))
+            {
+                return StatusCode(HttpStatusCode.Forbidden);
             }
 
             db.ExistingProducts.Remove(existingProduct);
