+++ SOPS/Controllers/ProductController.cs
@@ -10,6 +10,7 @@ using System.Web.Http;
 using System.Web.Http.Description;
 using Microsoft.AspNet.Identity;
 using SOPS.Attributes;
+using SOPS.ModelHelpers;
 using SOPS.Models;
 
 namespace SOPS.Controllers
@@ -40,6 +41,7 @@ namespace SOPS.Controllers
         }
 
         // PUT: api/Products/5
+        [Authorize(Roles = "Employee, Administrator")]
         [ResponseType(typeof(void))]
         public IHttpActionResult PutProduct(int id, Product product)
         {
@@ -51,6 +53,11 @@ namespace SOPS.Controllers
             if (id != product.Id)
             {
                 return BadRequest();
+            }
+
+            if (!db.IsCurrentUserEmployedInCompanyOrAdministrator(product.CompanyId))
+            {
+                return StatusCode(HttpStatusCode.Forbidden);
             }
 
             db.Entry(product).State = EntityState.Modified;
@@ -75,7 +82,7 @@ namespace SOPS.Controllers
         }
 
         // POST: api/Products
-        [Authorize]
+        [Authorize(Roles = "Employee, Administrator")]
         [ResponseType(typeof(Product))]
         public IHttpActionResult PostProduct(Product product)
         {
@@ -84,9 +91,7 @@ namespace SOPS.Controllers
                 return BadRequest(ModelState);
             }
 
-            var userId = User.Identity.GetUserId();
-            var asEmployee = db.Employees.Find(userId);
-            if(asEmployee == null || asEmployee.Company.Id != product.CompanyId)
+            if (!db.IsCurrentUserEmployedInCompanyOrAdministrator(product.CompanyId))
             {
                 return StatusCode(HttpStatusCode.Forbidden);
             }
@@ -98,6 +103,7 @@ namespace SOPS.Controllers
         }
 
         // DELETE: api/Products/5
+        [Authorize(Roles = "Employee, Administrator")]
         [ResponseType(typeof(Product))]
         public IHttpActionResult DeleteProduct(int id)
         {
@@ -105,6 +111,11 @@ namespace SOPS.Controllers
             if (product == null)
             {
                 return NotFound();
+            }
+
+            if (!db.IsCurrentUserEmployedInCompanyOrAdministrator(product.CompanyId))
+            {
+                return StatusCode(HttpStatusCode.Forbidden);
             }
 
             db.Products.Remove(product);
